+++ packages/@runlightyear/notion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/notion
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.28.0
+
 ## 0.1.9
 
 ### Patch Changes