+++ packages/@runlightyear/salesforce/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/salesforce
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.28.0
+
 ## 0.2.6
 
 ### Patch Changes